Bug 2020611: Improve performance of Jenkins tests
Restructure jenkins ext tests for origin/build speed up.
Cherry-pick of 6ade13f.
